+++ app/src/main/java/de/markusfisch/android/shadereditor/opengl/ShaderRenderer.java
@@ -1328,6 +1328,7 @@ public class ShaderRenderer implements GLSurfaceView.Renderer {
 	}
 
 	private String indexTextureNames(String source) {
+		InfoLog.resetSilentlyAddedExtraLines();
 		if (source == null) {
 			return null;
 		}
@@ -1394,10 +1395,9 @@ public class ShaderRenderer implements GLSurfaceView.Renderer {
 
 	private static String addPreprocessorDirective(String source,
 			String directive) {
-		boolean hasVersion = source.trim().startsWith("#version");
-		directive += "#line " + (hasVersion ? 1 : 0) + "\n";
+		InfoLog.addSilentlyAddedExtraLine();
 		// #version must always be the very first directive.
-		if (hasVersion) {
+		if (source.trim().startsWith("#version")) {
 			int lf = source.indexOf("\n");
 			if (lf < 0) {
 				// No line break?